+++ src/App.js
@@ -1,9 +1,6 @@
 import './languages';
-import messaging from '@react-native-firebase/messaging';
 import * as Sentry from '@sentry/react-native';
-import { QueryClientProvider } from '@tanstack/react-query';
 import { nanoid } from 'nanoid/non-secure';
-import PropTypes from 'prop-types';
 import React, { Component, createRef } from 'react';
 import {
   AppRegistry,
@@ -14,6 +11,7 @@ import {
   LogBox,
   View,
 } from 'react-native';
+
 // eslint-disable-next-line import/default
 import codePush from 'react-native-code-push';
 import {
@@ -27,7 +25,7 @@ import RNIOS11DeviceCheck from 'react-native-ios11-devicecheck';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { enableScreens } from 'react-native-screens';
 import VersionNumber from 'react-native-version-number';
-import { connect, Provider } from 'react-redux';
+import { connect, Provider as ReduxProvider } from 'react-redux';
 import { RecoilRoot } from 'recoil';
 import { runCampaignChecks } from './campaigns/campaignChecks';
 import PortalConsumer from './components/PortalConsumer';
@@ -52,7 +50,6 @@ import { isL2Network } from './handlers/web3';
 import RainbowContextWrapper from './helpers/RainbowContext';
 import isTestFlight from './helpers/isTestFlight';
 import networkTypes from './helpers/networkTypes';
-import { registerTokenRefreshListener, saveFCMToken } from './model/firebase';
 import * as keychain from './model/keychain';
 import { loadAddress } from './model/wallet';
 import { Navigation } from './navigation';
@@ -60,11 +57,14 @@ import RoutesComponent from './navigation/Routes';
 import { PerformanceContextMap } from './performance/PerformanceContextMap';
 import { PerformanceTracking } from './performance/tracking';
 import { PerformanceMetrics } from './performance/tracking/types/PerformanceMetrics';
-import { queryClient } from './react-query/queryClient';
+import {
+  PersistQueryClientProvider,
+  persistOptions,
+  queryClient,
+} from './react-query';
 import { additionalDataUpdateL2AssetBalance } from './redux/additionalAssetsData';
 import { explorerInitL2 } from './redux/explorer';
 import { fetchOnchainBalances } from './redux/fallbackExplorer';
-import { requestsForTopic } from './redux/requests';
 import store from './redux/store';
 import { uniswapPairsInit } from './redux/uniswap';
 import { walletConnectLoadState } from './redux/walletconnect';
@@ -81,8 +81,7 @@ import { InitialRouteContext } from '@/navigation/initialRoute';
 import Routes from '@/navigation/routesNames';
 import logger from '@/utils/logger';
 import { Portal } from '@/react-native-cool-modals/Portal';
-
-const WALLETCONNECT_SYNC_DELAY = 500;
+import { NotificationsHandler } from '@/notifications/NotificationsHandler';
 
 const FedoraToastRef = createRef();
 
@@ -146,10 +145,6 @@ enableScreens();
 const containerStyle = { flex: 1 };
 
 class App extends Component {
-  static propTypes = {
-    requestsForTopic: PropTypes.func,
-  };
-
   state = { appState: AppState.currentState, initialRoute: null };
 
   async componentDidMount() {
@@ -164,23 +159,7 @@ class App extends Component {
     rainbowTokenList.on('update', this.handleTokenListUpdate);
     appEvents.on('transactionConfirmed', this.handleTransactionConfirmed);
     await this.handleInitializeAnalytics();
-    saveFCMToken();
-    this.onTokenRefreshListener = registerTokenRefreshListener();
-
-    this.foregroundNotificationListener = messaging().onMessage(
-      this.onRemoteNotification
-    );
-
-    this.backgroundNotificationListener = messaging().setBackgroundMessageHandler(
-      async remoteMessage => {
-        setTimeout(() => {
-          this.onPushNotificationOpened(remoteMessage?.data?.topic);
-        }, WALLETCONNECT_SYNC_DELAY);
-      }
-    );
-
     this.branchListener = branchListener(this.handleOpenLinkingURL);
-
     // Walletconnect uses direct deeplinks
     if (android) {
       try {
@@ -222,9 +201,6 @@ class App extends Component {
   componentWillUnmount() {
     AppState.removeEventListener('change', this.handleAppStateChange);
     rainbowTokenList?.off?.('update', this.handleTokenListUpdate);
-    this.onTokenRefreshListener?.();
-    this.foregroundNotificationListener?.();
-    this.backgroundNotificationListener?.();
     this.branchListener?.();
   }
 
@@ -239,26 +215,8 @@ class App extends Component {
     store.dispatch(uniswapPairsInit());
   }
 
-  onRemoteNotification = notification => {
-    setTimeout(() => {
-      this.onPushNotificationOpened(notification?.data?.topic);
-    }, WALLETCONNECT_SYNC_DELAY);
-  };
-
   handleOpenLinkingURL = url => {
     handleDeeplink(url, this.state.initialRoute);
-  };
-
-  onPushNotificationOpened = topic => {
-    const { requestsForTopic } = this.props;
-    const requests = requestsForTopic(topic);
-    if (requests) {
-      // WC requests will open automatically
-      return false;
-    }
-    // In the future, here  is where we should
-    // handle all other kinds of push notifications
-    // For ex. incoming txs, etc.
   };
 
   handleInitializeAnalytics = async () => {
@@ -342,47 +300,51 @@ class App extends Component {
 
   render = () => (
     <MainThemeProvider>
-      <RainbowContextWrapper>
-        <ErrorBoundary>
-          <Portal>
-            <SafeAreaProvider>
-              <QueryClientProvider client={queryClient}>
-                <Provider store={store}>
+      <ReduxProvider store={store}>
+        <RainbowContextWrapper>
+          <ErrorBoundary>
+            <Portal>
+              <SafeAreaProvider>
+                <PersistQueryClientProvider
+                  client={queryClient}
+                  persistOptions={persistOptions}
+                >
                   <RecoilRoot>
                     <SharedValuesProvider>
-                      <View style={containerStyle}>
-                        {this.state.initialRoute && (
-                          <InitialRouteContext.Provider
-                            value={this.state.initialRoute}
-                          >
-                            <RoutesComponent
-                              onReady={this.handleSentryNavigationIntegration}
-                              ref={this.handleNavigatorRef}
-                            />
-                            <PortalConsumer />
-                          </InitialRouteContext.Provider>
-                        )}
-                        <OfflineToast />
-                        <FedoraToast ref={FedoraToastRef} />
-                      </View>
+                      <NotificationsHandler
+                        walletReady={this.props.walletReady}
+                      >
+                        <View style={containerStyle}>
+                          {this.state.initialRoute && (
+                            <InitialRouteContext.Provider
+                              value={this.state.initialRoute}
+                            >
+                              <RoutesComponent
+                                onReady={this.handleSentryNavigationIntegration}
+                                ref={this.handleNavigatorRef}
+                              />
+                              <PortalConsumer />
+                            </InitialRouteContext.Provider>
+                          )}
+                          <OfflineToast />
+                          <FedoraToast ref={FedoraToastRef} />
+                        </View>
+                      </NotificationsHandler>
                     </SharedValuesProvider>
                   </RecoilRoot>
-                </Provider>
-              </QueryClientProvider>
-            </SafeAreaProvider>
-          </Portal>
-        </ErrorBoundary>
-      </RainbowContextWrapper>
+                </PersistQueryClientProvider>
+              </SafeAreaProvider>
+            </Portal>
+          </ErrorBoundary>
+        </RainbowContextWrapper>
+      </ReduxProvider>
     </MainThemeProvider>
   );
 }
 
-const AppWithRedux = connect(
-  ({ appState: { walletReady } }) => ({ walletReady }),
-  {
-    requestsForTopic,
-  }
-)(App);
+const AppWithRedux = connect(({ appState: { walletReady } }) => ({
+  walletReady,
+}))(App);
 
 const AppWithReduxStore = () => <AppWithRedux store={store} />;
 
@@ -392,6 +354,12 @@ const codePushOptions = { checkFrequency: codePush.CheckFrequency.MANUAL };
 
 const AppWithCodePush = codePush(codePushOptions)(AppWithSentry);
 
+const PlaygroundWithReduxStore = () => (
+  <ReduxProvider store={store}>
+    <Playground />
+  </ReduxProvider>
+);
+
 AppRegistry.registerComponent('Rainbow', () =>
-  designSystemPlaygroundEnabled ? Playground : AppWithCodePush
+  designSystemPlaygroundEnabled ? PlaygroundWithReduxStore : AppWithCodePush
 );