+++ src/design-system/components/Box/Box.examples.tsx
@@ -14,9 +14,14 @@ export const background: Example = {
   Example: () =>
     source(
       <>
-        <Box background="surface" padding="20px">
+        <Box background="surfacePrimary" padding="20px">
           <Text color="label" size="17pt" weight="bold">
-            surface
+            surfacePrimary
+          </Text>
+        </Box>
+        <Box background="surfacePrimaryElevated" padding="20px">
+          <Text color="label" size="17pt" weight="bold">
+            surfacePrimaryElevated
           </Text>
         </Box>
         <Box background="surfaceSecondary" padding="20px">
@@ -24,9 +29,9 @@ export const background: Example = {
             surfaceSecondary
           </Text>
         </Box>
-        <Box background="surfaceTertiary" padding="20px">
+        <Box background="surfaceSecondaryElevated" padding="20px">
           <Text color="label" size="17pt" weight="bold">
-            surfaceTertiary
+            surfaceSecondaryElevated
           </Text>
         </Box>
         <Box background="fill" padding="20px">
@@ -37,11 +42,6 @@ export const background: Example = {
         <Box background="fillSecondary" padding="20px">
           <Text color="label" size="17pt" weight="bold">
             fillSecondary
-          </Text>
-        </Box>
-        <Box background="fillTertiary" padding="20px">
-          <Text color="label" size="17pt" weight="bold">
-            fillTertiary
           </Text>
         </Box>
         <Box background="accent" padding="20px">
@@ -108,25 +108,25 @@ export const padding: Example = {
   Example: () =>
     source(
       <Stack space="12px">
-        <Box background="surface" padding="20px">
+        <Box background="surfacePrimary" padding="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingHorizontal="20px">
+        <Box background="surfacePrimary" paddingHorizontal="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingVertical="20px">
+        <Box background="surfacePrimary" paddingVertical="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingLeft="20px">
+        <Box background="surfacePrimary" paddingLeft="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingRight="20px">
+        <Box background="surfacePrimary" paddingRight="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingTop="20px">
+        <Box background="surfacePrimary" paddingTop="20px">
           <Placeholder />
         </Box>
-        <Box background="surface" paddingBottom="20px">
+        <Box background="surfacePrimary" paddingBottom="20px">
           <Placeholder />
         </Box>
       </Stack>
@@ -138,31 +138,31 @@ export const margin: Example = {
   Example: () =>
     source(
       <Stack space="12px">
-        <Box background="surface" margin="-20px">
+        <Box background="surfacePrimary" margin="-20px">
           <Placeholder />
         </Box>
         <Inset vertical="20px">
-          <Box background="surface" marginHorizontal="-20px">
+          <Box background="surfacePrimary" marginHorizontal="-20px">
             <Placeholder />
           </Box>
         </Inset>
-        <Box background="surface" marginVertical="-20px">
+        <Box background="surfacePrimary" marginVertical="-20px">
           <Placeholder />
         </Box>
         <Inset vertical="20px">
-          <Box background="surface" marginLeft="-20px">
+          <Box background="surfacePrimary" marginLeft="-20px">
             <Placeholder />
           </Box>
         </Inset>
-        <Box background="surface" marginRight="-20px">
+        <Box background="surfacePrimary" marginRight="-20px">
           <Placeholder />
         </Box>
         <Inset vertical="20px">
-          <Box background="surface" marginTop="-20px">
+          <Box background="surfacePrimary" marginTop="-20px">
             <Placeholder />
           </Box>
         </Inset>
-        <Box background="surface" marginBottom="-20px">
+        <Box background="surfacePrimary" marginBottom="-20px">
           <Placeholder />
         </Box>
       </Stack>
@@ -217,7 +217,9 @@ export const heights: Example = {
 export const shadows: Example = {
   name: 'Shadows',
   Example: () =>
-    source(<Box background="surface" padding="24px" shadow="30px light" />),
+    source(
+      <Box background="surfacePrimary" padding="24px" shadow="30px light" />
+    ),
 };
 
 export const shadowsWithSizes: Example = {
@@ -229,31 +231,55 @@ export const shadowsWithSizes: Example = {
         <Columns space="32px">
           <Column width="1/3" />
           <Column width="1/3">
-            <Box background="surface" padding="16px" shadow="9px medium" />
+            <Box
+              background="surfacePrimary"
+              padding="16px"
+              shadow="9px medium"
+            />
           </Column>
         </Columns>
         <Columns space="32px">
           <Column width="1/3" />
           <Column width="1/3">
-            <Box background="surface" padding="16px" shadow="12px medium" />
+            <Box
+              background="surfacePrimary"
+              padding="16px"
+              shadow="12px medium"
+            />
           </Column>
           <Column width="1/3">
-            <Box background="surface" padding="16px" shadow="12px heavy" />
+            <Box
+              background="surfacePrimary"
+              padding="16px"
+              shadow="12px heavy"
+            />
           </Column>
         </Columns>
         <Columns space="32px">
           <Column width="1/3">
-            <Box background="surface" padding="16px" shadow="21px light" />
+            <Box
+              background="surfacePrimary"
+              padding="16px"
+              shadow="21px light"
+            />
           </Column>
           <Column width="1/3" />
           <Column width="1/3">
-            <Box background="surface" padding="16px" shadow="21px heavy" />
+            <Box
+              background="surfacePrimary"
+              padding="16px"
+              shadow="21px heavy"
+            />
           </Column>
         </Columns>
         <Columns space="32px">
-          <Box background="surface" padding="16px" shadow="30px light" />
-          <Box background="surface" padding="16px" shadow="30px medium" />
-          <Box background="surface" padding="16px" shadow="30px heavy" />
+          <Box background="surfacePrimary" padding="16px" shadow="30px light" />
+          <Box
+            background="surfacePrimary"
+            padding="16px"
+            shadow="30px medium"
+          />
+          <Box background="surfacePrimary" padding="16px" shadow="30px heavy" />
         </Columns>
       </Stack>
     ),
@@ -265,14 +291,18 @@ export const shadowsWithColors: Example = {
   Example: () =>
     source(
       <Columns space="32px">
-        <Box background="surface" padding="16px" shadow="12px medium accent" />
         <Box
-          background="surface"
+          background="surfacePrimary"
+          padding="16px"
+          shadow="12px medium accent"
+        />
+        <Box
+          background="surfacePrimary"
           padding="16px"
           shadow="21px heavy swap (Deprecated)"
         />
         <Box
-          background="surface"
+          background="surfacePrimary"
           padding="16px"
           shadow="30px heavy action (Deprecated)"
         />
@@ -287,7 +317,7 @@ export const shadowsWithCustom: Example = {
     source(
       <Columns space="32px">
         <Box
-          background="surface"
+          background="surfacePrimary"
           padding="16px"
           shadow={{
             custom: {
@@ -311,7 +341,7 @@ export const shadowsWithCustom: Example = {
           }}
         />
         <Box
-          background="surface"
+          background="surfacePrimary"
           padding="16px"
           shadow={{
             custom: {